==> camels_de/version1/scripts/03_humaninfluence_attributes_harz.ipynb <==
# Climatic attributes

Notebook to create the file `CAMELS_DE_humaninfluence_attributes.csv`.  

columns:
- gauge_id
- dams_names
- dams_river_names
- dams_num
- dams_year_first
- dams_year_last
- dams_total_lake_area [km²]
- dams_total_lake_volume [Mio. m³]
- dams_purposes

import os
from glob import glob
import pandas as pd

# get camels_ids from hydromet timeseries
camels_ids = [camels_id.split("_")[-1].split(".csv")[0] for camels_id in glob("../output_data/camels_de/timeseries/*.csv")]

# sort camels_ids
camels_ids = sorted(camels_ids)

print(f"Total number of stations in CAMELS-DE v1: {len(camels_ids)}")

## Read human influence data

### Inventory of dams in Germany (DIG)

We can use the data as is.

# path of extracted dam data
path = os.path.join('..', '..', '..', 'camels_human_influence', 'dams', 'output_data', 'dams_in_germany.csv')

# read dam data
dam_data = pd.read_csv(path)

# filter for camels_ids and sort by camels_id
dam_data = dam_data.sort_values("camels_id").reset_index(drop=True)

# make columns dams_year_last and dams_year_first integer
dam_data["dams_year_first"] = dam_data["dams_year_first"].astype("Int64")
dam_data["dams_year_last"] = dam_data["dams_year_last"].astype("Int64")

# rename column camels_id to gauge_id
dam_data = dam_data.rename(columns={"camels_id": "gauge_id"})

dam_data

# save human influences data
dam_data.to_csv("../output_data/camels_de/CAMELS_DE_humaninfluence_attributes.csv", index=False)
