==> camels_de/version1/scripts/05_soil_attributes_harz.ipynb <==
# Soil attributes

Notebook to create the file `CAMELS_DE_soil_attributes.csv`.  

columns:
- gauge_id
- **clay_0_30cm_mean [g/100g (%)]**
- clay_0_30cm_min [g/100g (%)]
- clay_0_30cm_max [g/100g (%)]
- clay_0_30cm_q05 [g/100g (%)]
- clay_0_30cm_q50 [g/100g (%)]
- clay_0_30cm_q95 [g/100g (%)]
- clay_0_30cm_stdev [g/100g (%)]
- clay_30_100cm_mean [g/100g (%)]
- clay_30_100cm_min [g/100g (%)]
- clay_30_100cm_max [g/100g (%)]
- clay_30_100cm_q05 [g/100g (%)]
- clay_30_100cm_q50 [g/100g (%)]
- clay_30_100cm_q95 [g/100g (%)]
- clay_30_100cm_stdev [g/100g (%)]
- clay_100_200cm_mean [g/100g (%)]
- clay_100_200cm_min [g/100g (%)]
- clay_100_200cm_max [g/100g (%)]
- clay_100_200cm_q05 [g/100g (%)]
- clay_100_200cm_q50 [g/100g (%)]
- clay_100_200cm_q95 [g/100g (%)]
- clay_100_200cm_stdev [g/100g (%)]
- **sand_0_30cm_mean [g/100g (%)]**
- sand_0_30cm_min [g/100g (%)]
- sand_0_30cm_max [g/100g (%)]
- sand_0_30cm_q05 [g/100g (%)]
- sand_0_30cm_q50 [g/100g (%)]
- sand_0_30cm_q95 [g/100g (%)]
- sand_0_30cm_stdev [g/100g (%)]
- sand_30_100cm_mean [g/100g (%)]
- sand_30_100cm_min [g/100g (%)]
- sand_30_100cm_max [g/100g (%)]
- sand_30_100cm_q05 [g/100g (%)]
- sand_30_100cm_q50 [g/100g (%)]
- sand_30_100cm_q95 [g/100g (%)]
- sand_30_100cm_stdev [g/100g (%)]
- sand_100_200cm_mean [g/100g (%)]
- sand_100_200cm_min [g/100g (%)]
- sand_100_200cm_max [g/100g (%)]
- sand_100_200cm_q05 [g/100g (%)]
- sand_100_200cm_q50 [g/100g (%)]
- sand_100_200cm_q95 [g/100g (%)]
- sand_100_200cm_stdev [g/100g (%)]
- **silt_0_30cm_mean [g/100g (%)]**
- silt_0_30cm_min [g/100g (%)]
- silt_0_30cm_max [g/100g (%)]
- silt_0_30cm_q05 [g/100g (%)]
- silt_0_30cm_q50 [g/100g (%)]
- silt_0_30cm_q95 [g/100g (%)]
- silt_0_30cm_stdev [g/100g (%)]
- silt_30_100cm_mean [g/100g (%)]
- silt_30_100cm_min [g/100g (%)]
- silt_30_100cm_max [g/100g (%)]
- silt_30_100cm_q05 [g/100g (%)]
- silt_30_100cm_q50 [g/100g (%)]
- silt_30_100cm_q95 [g/100g (%)]
- silt_30_100cm_stdev [g/100g (%)]
- silt_100_200cm_mean [g/100g (%)]
- silt_100_200cm_min [g/100g (%)]
- silt_100_200cm_max [g/100g (%)]
- silt_100_200cm_q05 [g/100g (%)]
- silt_100_200cm_q50 [g/100g (%)]
- silt_100_200cm_q95 [g/100g (%)]
- silt_100_200cm_stdev [g/100g (%)]
- **coarse_fragments_0_30cm_mean [cm³/100cm³ (vol%)]**
- coarse_fragments_0_30cm_min [cm³/100cm³ (vol%)]
- coarse_fragments_0_30cm_max [cm³/100cm³ (vol%)]
- coarse_fragments_0_30cm_q05 [cm³/100cm³ (vol%)]
- coarse_fragments_0_30cm_q50 [cm³/100cm³ (vol%)]
- coarse_fragments_0_30cm_q95 [cm³/100cm³ (vol%)]
- coarse_fragments_0_30cm_stdev [cm³/100cm³ (vol%)]
- coarse_fragments_30_100cm_mean [cm³/100cm³ (vol%)]
- coarse_fragments_30_100cm_min [cm³/100cm³ (vol%)]
- coarse_fragments_30_100cm_max [cm³/100cm³ (vol%)]
- coarse_fragments_30_100cm_q05 [cm³/100cm³ (vol%)]
- coarse_fragments_30_100cm_q50 [cm³/100cm³ (vol%)]
- coarse_fragments_30_100cm_q95 [cm³/100cm³ (vol%)]
- coarse_fragments_30_100cm_stdev [cm³/100cm³ (vol%)]
- coarse_fragments_100_200cm_mean [cm³/100cm³ (vol%)]
- coarse_fragments_100_200cm_min [cm³/100cm³ (vol%)]
- coarse_fragments_100_200cm_max [cm³/100cm³ (vol%)]
- coarse_fragments_100_200cm_q05 [cm³/100cm³ (vol%)]
- coarse_fragments_100_200cm_q50 [cm³/100cm³ (vol%)]
- coarse_fragments_100_200cm_q95 [cm³/100cm³ (vol%)]
- coarse_fragments_100_200cm_stdev [cm³/100cm³ (vol%)]
- **soil_organic_carbon_0_30cm_mean [g/kg]**
- soil_organic_carbon_0_30cm_min [g/kg]
- soil_organic_carbon_0_30cm_max [g/kg]
- soil_organic_carbon_0_30cm_q05 [g/kg]
- soil_organic_carbon_0_30cm_q50 [g/kg]
- soil_organic_carbon_0_30cm_q95 [g/kg]
- soil_organic_carbon_0_30cm_stdev [g/kg]
- soil_organic_carbon_30_100cm_mean [g/kg]
- soil_organic_carbon_30_100cm_min [g/kg]
- soil_organic_carbon_30_100cm_max [g/kg]
- soil_organic_carbon_30_100cm_q05 [g/kg]
- soil_organic_carbon_30_100cm_q50 [g/kg]
- soil_organic_carbon_30_100cm_q95 [g/kg]
- soil_organic_carbon_30_100cm_stdev [g/kg]
- soil_organic_carbon_100_200cm_mean [g/kg]
- soil_organic_carbon_100_200cm_min [g/kg]
- soil_organic_carbon_100_200cm_max [g/kg]
- soil_organic_carbon_100_200cm_q05 [g/kg]
- soil_organic_carbon_100_200cm_q50 [g/kg]
- soil_organic_carbon_100_200cm_q95 [g/kg]
- soil_organic_carbon_100_200cm_stdev [g/kg]
- **bulk_density_0_30cm_mean [kg/dm³]**
- bulk_density_0_30cm_min [kg/dm³]
- bulk_density_0_30cm_max [kg/dm³]
- bulk_density_0_30cm_q05 [kg/dm³]
- bulk_density_0_30cm_q50 [kg/dm³]
- bulk_density_0_30cm_q95 [kg/dm³]
- bulk_density_0_30cm_stdev [kg/dm³]
- bulk_density_30_100cm_mean [kg/dm³]
- bulk_density_30_100cm_min [kg/dm³]
- bulk_density_30_100cm_max [kg/dm³]
- bulk_density_30_100cm_q05 [kg/dm³]
- bulk_density_30_100cm_q50 [kg/dm³]
- bulk_density_30_100cm_q95 [kg/dm³]
- bulk_density_30_100cm_stdev [kg/dm³]
- bulk_density_100_200cm_mean [kg/dm³]
- bulk_density_100_200cm_min [kg/dm³]
- bulk_density_100_200cm_max [kg/dm³]
- bulk_density_100_200cm_q05 [kg/dm³]
- bulk_density_100_200cm_q50 [kg/dm³]
- bulk_density_100_200cm_q95 [kg/dm³]
- bulk_density_100_200cm_stdev [kg/dm³]


import os
from glob import glob
import pandas as pd

# get camels_ids from hydromet timeseries
camels_ids = [camels_id.split("_")[-1].split(".csv")[0] for camels_id in glob("../output_data/camels_de/timeseries/*.csv")]

# sort camels_ids
camels_ids = sorted(camels_ids)

print(f"Total number of stations in CAMELS-DE v1: {len(camels_ids)}")

## Read and save soil data

We extracted the hydrogeology attributes from the ISRIC soil grid dataset, read, merge and process here.

# get all soil grids variables
variables = ["clay", "silt", "sand", "cfvo", "bdod", "soc"]
# get all files within folder
files = [f'../../../camels_soils/isric/output_data/isric_processed/{variable}_processed.csv'for variable in variables]
soil_data = pd.DataFrame(columns=["camels_id"], data=camels_ids)

# read files and merge, join on camels_id
for file in files:
    # read file
    data = pd.read_csv(file)

    # turn camels_id into object
    data["camels_id"] = data["camels_id"].astype(str)

    # merge
    soil_data = soil_data.merge(data, how="left", on="camels_id")

# filter for camels_ids and sort by camels_id
soil_data = soil_data[soil_data["camels_id"].isin(camels_ids)].sort_values("camels_id").reset_index(drop=True)

# rename column camels_id to gauge_id
soil_data = soil_data.rename(columns={"camels_id": "gauge_id"})

# remove all columns that do not have "mean" in the name
soil_data = soil_data[[col for col in soil_data.columns if "mean" in col or col == "gauge_id"]]

# round to 2 decimal places
soil_data = soil_data.round(2)

# save to csv
soil_data.to_csv("../output_data/camels_de/CAMELS_DE_soil_attributes.csv", index=False)
