==> camels_de/version1/scripts/06_landcover_attributes_harz.ipynb <==
# Landcover attributes

Notebook to create the file `CAMELS_DE_landcover_attributes.csv`.

columns:

- gauge_id
- artificial_surfaces_perc [%]
- agricultural_areas_perc [%]
- forests_and_seminatural_areas_perc [%]
- wetlands_perc [%]
- water_bodies_perc [%]


import os
from glob import glob
import pandas as pd

# get camels_ids from hydromet timeseries
camels_ids = [camels_id.split("_")[-1].split(".csv")[0] for camels_id in glob("../output_data/camels_de/timeseries/*.csv")]

# sort camels_ids
camels_ids = sorted(camels_ids)

print(f"Total number of stations in CAMELS-DE v1: {len(camels_ids)}")

## Read landcover data

We extracted the landcover attributes from the CORINE dataset, read and process here.


# read landcover data
df = pd.read_csv(os.path.join('../../..', "camels_landcover/corine/output_data/corine_extracted.csv"))

# turn camels_id into object
df["camels_id"] = df["camels_id"].astype(str)

# filter for camels_ids and sort by camels_id
df = df[df["camels_id"].isin(camels_ids)].sort_values("camels_id").reset_index(drop=True)

# round to 2 decimal places
df = df.round(2)

# rename column camels_id to gauge_id
df = df.rename(columns={"camels_id": "gauge_id"})

df.to_csv("../output_data/camels_de/CAMELS_DE_landcover_attributes.csv", index=False)
