==> camels_de/version1/scripts/07_topographic_attributes.ipynb <==
# Topographic attributes

Notebook to create the file `CAMELS_DE_topographic_attributes.csv`.  

columns in CAMELS-GB:
- gauge_id
- provider_id
- gauge_name
- water_body_name
- federal_state
- gauge_lat [°]
- gauge_lon [°]
- gauge_easting [m]
- gauge_northing [m]
- gauge_elev_metadata [m a.s.l.]
- area_metadata [km²]
- gauge_elev [m a.s.l.]
- area [km²]
- elev_mean [m a.s.l.]
- elev_min [m a.s.l.]
- elev_5 [m a.s.l.]
- elev_50 [m a.s.l.]
- elev_95 [m a.s.l.]
- elev_max [m a.s.l.]


from glob import glob
import pandas as pd

from camelsp import get_metadata, Station

# get camels_ids from hydromet timeseries
camels_ids = [camels_id.split("_")[-1].split(".csv")[0] for camels_id in glob("../output_data/camels_de/timeseries/*.csv")]

# sort camels_ids
camels_ids = sorted(camels_ids)

print(f"Total number of stations in CAMELS-DE v1: {len(camels_ids)}")

## Read DEM data

# read metadata, filter for camels_ids
meta = get_metadata()
meta = meta[meta["camels_id"].isin(camels_ids)].sort_values("camels_id").reset_index(drop=True)


dem = pd.read_csv("../../../camelsp/output_data/raw_catchment_attributes/dem/copernicus_dem30/dem_extracted.csv")
dem = dem[dem["camels_id"].isin(camels_ids)].sort_values("camels_id").reset_index(drop=True)

dem

## Create CAMELS-DE topographic attributes csv

Rename columns to be consistent with CAMELS-GB, CAMELS-CH

# dataframe to store results
df_results = pd.DataFrame(index=camels_ids)

for camels_id in camels_ids:
    # add provider_id
    provider_id = meta[meta["camels_id"] == camels_id]["provider_id"].values[0]
    df_results.loc[camels_id, "provider_id"] = provider_id

    # add gauge_name
    gauge_name = meta[meta["camels_id"] == camels_id]["gauge_name"].values[0]
    df_results.loc[camels_id, "gauge_name"] = gauge_name

    # add water_body_name
    water_body_name = meta[meta["camels_id"] == camels_id]["waterbody_name"].values[0]
    df_results.loc[camels_id, "water_body_name"] = water_body_name

    # add federal_state
    federal_state = meta[meta["camels_id"] == camels_id]["federal_state"].values[0]
    df_results.loc[camels_id, "federal_state"] = federal_state

    # add latitude and longitude
    gauge_lat = meta[meta["camels_id"] == camels_id]["lat"].values[0]
    gauge_lon = meta[meta["camels_id"] == camels_id]["lon"].values[0]
    df_results.loc[camels_id, "gauge_lat"] = round(gauge_lat, 6) # round to 6 decimal places to have a precision of ~0.11m
    df_results.loc[camels_id, "gauge_lon"] = round(gauge_lon, 6) # round to 6 decimal places to have a precision of ~0.11m

    # add easting and northing as integers
    gauge_easting = meta[meta["camels_id"] == camels_id]["x"].values[0]
    gauge_northing = meta[meta["camels_id"] == camels_id]["y"].values[0]
    df_results.loc[camels_id, "gauge_easting"] = int(gauge_easting)
    df_results.loc[camels_id, "gauge_northing"] = int(gauge_northing)

    # add gauge elevation from metadata
    gauge_elev_metadata = meta[meta["camels_id"] == camels_id]["gauge_elevation"].values[0]
    df_results.loc[camels_id, "gauge_elev_metadata"] = round(gauge_elev_metadata, 2)

    # add area from metadata
    area_metadata = meta[meta["camels_id"] == camels_id]["area"].values[0]
    df_results.loc[camels_id, "area_metadata"] = round(area_metadata, 2)

    # add gauge elevation from dem
    gauge_elev = dem[meta["camels_id"] == camels_id]["gauge_elevation_from_dem"].values[0]
    df_results.loc[camels_id, "gauge_elev"] = round(gauge_elev, 2)

    # add merit hydro area
    s = Station(camels_id)
    area = (s.get_catchment("merit_hydro").to_crs("EPSG:6933").area / 1e6)[0]
    df_results.loc[camels_id, "area"] = round(area, 2)

    # add catchment elevation statistics
    elev_mean = dem[meta["camels_id"] == camels_id]["mean_catchment_elevation_from_dem"].values[0]
    elev_min = dem[meta["camels_id"] == camels_id]["min_catchment_elevation_from_dem"].values[0]
    elev_5 = dem[meta["camels_id"] == camels_id]["quantile0.05_catchment_elevation_from_dem"].values[0] 
    elev_50 = dem[meta["camels_id"] == camels_id]["quantile0.5_catchment_elevation_from_dem"].values[0]
    elev_95 = dem[meta["camels_id"] == camels_id]["quantile0.95_catchment_elevation_from_dem"].values[0]
    elev_max = dem[meta["camels_id"] == camels_id]["max_catchment_elevation_from_dem"].values[0]
    df_results.loc[camels_id, "elev_mean"] = round(elev_mean, 2)
    df_results.loc[camels_id, "elev_min"] = round(elev_min, 2)
    df_results.loc[camels_id, "elev_5"] = round(elev_5, 2)
    df_results.loc[camels_id, "elev_50"] = round(elev_50, 2)
    df_results.loc[camels_id, "elev_95"] = round(elev_95, 2)
    df_results.loc[camels_id, "elev_max"] = round(elev_max, 2)

df_results

## Save results

df_results.to_csv("../output_data/camels_de/CAMELS_DE_topographic_attributes.csv", index_label="gauge_id")
